Move redesign into v2 and paginate rankings
