+++ front-end/src/Routers/MainRouter.js
@@ -1,29 +1,36 @@
-import { BrowserRouter, Route, Routes } from 'react-router-dom';
+import { BrowserRouter, Route, Routes, useLocation } from 'react-router-dom';
 import Header from '../components/Header';
-import V2ReportMatch from '../components/ReportMatch';
-import V2LandingPage from '../components/LandingPage';
-import V2ResultsPage from '../components/Results';
-import V2Players from '../components/Players';
-import V2EloPage from '../components/EloPage';
-import PlayerProfile from '../components/PlayerProfile';
-import LegacyLandingPage from '../components/legacy/LegacyLandingPage';
-import LegacyReportMatch from '../components/legacy/LegacyReportMatch';
-import LegacyResults from '../components/legacy/LegacyResults';
-import LegacyPlayers from '../components/legacy/LegacyPlayers';
-import LegacyEloPage from '../components/legacy/LegacyEloPage';
+import LandingPage from '../components/LandingPage';
+import ReportMatchComponent from '../components/ReportMatch';
+import ResultsPage from '../components/Results';
+import Players from '../components/Players';
+import EloPage from '../components/EloPage';
+import V2Header from '../v2/components/Header';
+import V2ReportMatch from '../v2/components/ReportMatch';
+import V2LandingPage from '../v2/components/LandingPage';
+import V2ResultsPage from '../v2/components/Results';
+import V2Players from '../v2/components/Players';
+import V2EloPage from '../v2/components/EloPage';
+import PlayerProfile from '../v2/components/PlayerProfile';
 
+const RouteHeader = () => {
+    const location = useLocation();
+    const isV2 = location.pathname === '/v2' || location.pathname.startsWith('/v2/');
+
+    return isV2 ? <V2Header /> : <Header />;
+};
 
 const MainRouter = () => {
     return (
         <>
             <BrowserRouter>
-                <Header />
+                <RouteHeader />
                 <Routes>
-                    <Route exact path='/' element={<LegacyLandingPage />} />
-                    <Route exact path='/report' element={<LegacyReportMatch />} />
-                    <Route exact path='/results' element={<LegacyResults />} />
-                    <Route exact path='/players' element={<LegacyPlayers />} />
-                    <Route exact path='/elo' element={<LegacyEloPage />} />
+                    <Route exact path='/' element={<LandingPage />} />
+                    <Route exact path='/report' element={<ReportMatchComponent />} />
+                    <Route exact path='/results' element={<ResultsPage />} />
+                    <Route exact path='/players' element={<Players />} />
+                    <Route exact path='/elo' element={<EloPage />} />
                     <Route exact path='/v2' element={<V2LandingPage />} />
                     <Route exact path='/v2/report' element={<V2ReportMatch />} />
                     <Route exact path='/v2/results' element={<V2ResultsPage />} />
