+++ front-end/src/__tests__/v2/components/PlayerProfile.test.js
@@ -1,8 +1,8 @@
 import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { MemoryRouter, Route, Routes, useLocation } from 'react-router-dom';
-import { AppContext } from '../../Contexts/AppContext';
-import PlayerProfile from '../../components/PlayerProfile';
+import { AppContext } from '../../../Contexts/AppContext';
+import PlayerProfile from '../../../v2/components/PlayerProfile';
 
 const statsRecord = (id, name, wins, overrides = {}) => ({
   id,
