+++ front-end/src/__tests__/v2/components/Players.test.js
@@ -1,8 +1,8 @@
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { MemoryRouter } from 'react-router-dom';
-import { AppContext } from '../../Contexts/AppContext';
-import Players from '../../components/Players';
+import { AppContext } from '../../../Contexts/AppContext';
+import Players from '../../../v2/components/Players';
 
 const defaultPlayers = [
   { id: 113, first_name: 't_Liam', last_name: 'Zhang' },
