+++ front-end/src/components/EloPage.js
@@ -1,76 +1,123 @@
+import { Container, Row, Col, ListGroup } from "react-bootstrap";
 import { useContext, useEffect, useState } from "react";
-import { Link } from "react-router-dom";
 import { AppContext } from "../Contexts/AppContext";
-import { PageShell, RankingsTable, SeasonSelect, TabControl } from "./RedesignUI";
-import {
-    buildEloRankings,
-    cleanSeedPrefix,
-    SEASON_OPTIONS,
-    findPlayerByName,
-    getPlayerProfilePath,
-    parseSeasonValue,
-} from "../utils/playerViewModels";
-
-const ELO_EVENT_TABS = [
-    { key: 'singles', label: 'Singles' },
-    { key: 'doubles', label: 'Doubles' },
-];
+import SeasonSelector from "./SeasonSelector";
 
 const EloPage = () => {
-    const { players, queryElo } = useContext(AppContext);
-    const [seasonValue, setSeasonValue] = useState(SEASON_OPTIONS[0].value);
+    const { queryElo, players } = useContext(AppContext);
     const [seasonStart, setSeasonStart] = useState('2025-09-01');
     const [seasonEnd, setSeasonEnd] = useState('2026-08-31');
-    const [activeEvent, setActiveEvent] = useState('singles');
     const [singlesElo, setSinglesElo] = useState([]);
     const [doublesElo, setDoublesElo] = useState([]);
 
     useEffect(() => {
-        queryElo('singles', seasonStart, seasonEnd).then(data => setSinglesElo(data || []));
-        queryElo('doubles', seasonStart, seasonEnd).then(data => setDoublesElo(data || []));
-    }, [seasonStart, seasonEnd, queryElo])
+        queryElo('singles', seasonStart, seasonEnd).then(data => {
+            setSinglesElo(data);
+        });
+        queryElo('doubles', seasonStart, seasonEnd).then(data => {
+            setDoublesElo(data);
+        });
+    }, [queryElo, seasonStart, seasonEnd]);
 
-    const handleSeasonChange = (value, parsedSeason) => {
-        const season = parsedSeason || parseSeasonValue(value);
-        setSeasonValue(value);
-        setSeasonStart(season.start);
-        setSeasonEnd(season.end);
+    const doubles = (elo) => {
+        const result = [...elo]
+            .map(player => ({
+                ...player,
+                mu: player.doubles_rating.mu,
+                sigma: player.doubles_rating.sigma,
+            }))
+            .sort((a, b) => b.mu - a.mu)
+            .filter(player => player.doubles_games_played > 5 && player.sigma < 4);
+
+        return (
+            <ListGroup as="ol" numbered>
+                {result.map((player, index) => (
+                    <ListGroup.Item key={player.name}>
+                        <Row>
+                            <Col xs={5}>{index + 1}. {player.name}</Col>
+                            <Col xs={3}>{player.mu.toFixed(3)}</Col>
+                            <Col xs={2}>{player.sigma.toFixed(3)}</Col>
+                            <Col xs={2}>{player.doubles_win_pct.toFixed(3)}</Col>
+                        </Row>
+                    </ListGroup.Item>
+                ))}
+            </ListGroup>
+        );
     };
 
-    const rankings = buildEloRankings(activeEvent === 'singles' ? singlesElo : doublesElo, activeEvent)
-        .map(row => {
-            const localPlayer = findPlayerByName(players, row.name);
-            return {
-                ...row,
-                id: row.id || (localPlayer && localPlayer.id),
-                name: cleanSeedPrefix(row.name),
-                to: row.to || getPlayerProfilePath(localPlayer),
-            };
-        });
+    const singles = (elo) => {
+        const result = [...elo]
+            .map(player => ({
+                ...player,
+                mu: player.singles_rating.mu,
+                sigma: player.singles_rating.sigma,
+            }))
+            .sort((a, b) => b.mu - a.mu)
+            .filter(player => player.singles_games_played > 5 && player.sigma < 4);
 
-    const columns = [
-        { key: 'rank', label: 'Rank' },
-        {
-            key: 'name',
-            label: 'Name',
-            render: row => row.to
-                ? <Link className="table-name-link" to={row.to}>{row.name}</Link>
-                : row.name,
-        },
-        { key: 'mu', label: '\u03bc', render: row => Number.isFinite(row.mu) ? row.mu.toFixed(2) : row.mu },
-        { key: 'sigma', label: '\u03c3', render: row => Number.isFinite(row.sigma) ? row.sigma.toFixed(2) : row.sigma },
-        { key: 'winPct', label: 'Win %', render: row => `${row.winPct}%` },
-    ];
+        return (
+            <ListGroup as="ol" numbered>
+                {result.map((player, index) => (
+                    <ListGroup.Item key={player.name}>
+                        <Row>
+                            <Col xs={5}>{index + 1}. {player.name}</Col>
+                            <Col xs={3}>{player.mu.toFixed(3)}</Col>
+                            <Col xs={2}>{player.sigma.toFixed(3)}</Col>
+                            <Col xs={2}>{player.singles_win_pct.toFixed(3)}</Col>
+                        </Row>
+                    </ListGroup.Item>
+                ))}
+            </ListGroup>
+        );
+    };
 
     return (
-        <PageShell title="Rankings" eyebrow="Current elo rankings">
-            <div className="toolbar-row">
-                <TabControl tabs={ELO_EVENT_TABS} active={activeEvent} onChange={setActiveEvent} />
-                <SeasonSelect value={seasonValue} onChange={handleSeasonChange} />
-            </div>
-            <RankingsTable columns={columns} rows={rankings.slice(0, 10)} />
-        </PageShell>
-    )
-}
+        <>
+            <Container>
+                <Row className="page-title">
+                    Current Elo Rankings:
+                </Row>
+                <SeasonSelector
+                    setStart={(start) => setSeasonStart(start)}
+                    setEnd={(end) => setSeasonEnd(end)}
+                />
+                {(singlesElo.length === 0 || doublesElo.length === 0 || players.length === 0) ? (
+                    <Col className="page-title">
+                        Retreiving data, please be patient...
+                    </Col>
+                ) : (
+                    <Container>
+                        <Row>
+                            <Col sm={12} md={12}>
+                                <div className="table-header">Singles Elo</div>
+                                <ListGroup.Item>
+                                    <Row>
+                                        <Col xs={5}><b>Name</b></Col>
+                                        <Col xs={3}><b>{'\u03BC'}</b></Col>
+                                        <Col xs={2}><b>{'\u03C3'}</b></Col>
+                                        <Col xs={2}><b>Win %</b></Col>
+                                    </Row>
+                                </ListGroup.Item>
+                                {singles(singlesElo)}
+                            </Col>
+                            <Col sm={12} md={12}>
+                                <div className="table-header">Doubles Elo</div>
+                                <ListGroup.Item>
+                                    <Row>
+                                        <Col xs={5}><b>Name</b></Col>
+                                        <Col xs={3}><b>{'\u03BC'}</b></Col>
+                                        <Col xs={2}><b>{'\u03C3'}</b></Col>
+                                        <Col xs={2}><b>Win %</b></Col>
+                                    </Row>
+                                </ListGroup.Item>
+                                {doubles(doublesElo)}
+                            </Col>
+                        </Row>
+                    </Container>
+                )}
+            </Container>
+        </>
+    );
+};
 
 export default EloPage;
