+++ front-end/src/components/LandingPage.js
@@ -1,57 +1,100 @@
-import { useState, useEffect, useContext } from "react"
-import { Link } from "react-router-dom";
+import { Container, Row, Col, ListGroup } from "react-bootstrap";
+import { useState, useEffect, useContext } from "react";
 import { AppContext } from "../Contexts/AppContext";
-import { PageShell, RankingsTable, SeasonSelect, TabControl } from "./RedesignUI";
-import { buildWinRankings, EVENT_TABS, SEASON_OPTIONS, parseSeasonValue } from "../utils/playerViewModels";
+import SeasonSelector from "./SeasonSelector";
 
 const LandingPage = () => {
     const { queryStats } = useContext(AppContext);
     const [stats, setStats] = useState([]);
-    const [activeEvent, setActiveEvent] = useState('singles');
-    const [seasonValue, setSeasonValue] = useState(SEASON_OPTIONS[0].value);
+    const [singlesRankings, setSinglesRankings] = useState([]);
+    const [doublesRankings, setDoublesRankings] = useState([]);
+    const [mixedRankings, setMixedRankings] = useState([]);
     const [seasonStart, setSeasonStart] = useState('2025-09-01');
     const [seasonEnd, setSeasonEnd] = useState('2026-08-31');
 
-    const getStats = (newSeasonStart, newSeasonEnd) => {
-        queryStats(newSeasonStart, newSeasonEnd).then(data => {
-            setStats(data)
-        })
-    }
+    const updateStats = (category, newStats) => {
+        const results = newStats.map(s => {
+            let percentage = 0;
+            const categoryWins = `${category}_wins`;
+            const categoryLosses = `${category}_losses`;
+            if (s[categoryWins] + s[categoryLosses] !== 0) {
+                percentage = Math.round((s[categoryWins] / (s[categoryWins] + s[categoryLosses])) * 100);
+            }
+            return {
+                name: s.name,
+                percentage,
+                wins: s[categoryWins],
+                losses: s[categoryLosses],
+            };
+        });
+        const items = Object.keys(results).map(key => [key, results[key]]);
+        items.sort((first, second) => second[1].percentage - first[1].percentage);
+        if (category === 'singles') {
+            setSinglesRankings(items);
+        } else if (category === 'doubles') {
+            setDoublesRankings(items);
+        } else {
+            setMixedRankings(items);
+        }
+    };
 
     useEffect(() => {
-        getStats(seasonStart, seasonEnd);
-    }, [seasonStart, seasonEnd])
+        queryStats(seasonStart, seasonEnd).then(data => {
+            setStats(data);
+        });
+    }, [queryStats, seasonStart, seasonEnd]);
 
-    const handleSeasonChange = (value, parsedSeason) => {
-        const season = parsedSeason || parseSeasonValue(value);
-        setSeasonValue(value);
-        setSeasonStart(season.start);
-        setSeasonEnd(season.end);
-    }
+    useEffect(() => {
+        updateStats('singles', stats);
+        updateStats('doubles', stats);
+        updateStats('mixed', stats);
+    }, [stats]);
 
-    const rankings = buildWinRankings(stats, activeEvent).slice(0, 10);
-    const columns = [
-        { key: 'rank', label: 'Rank' },
-        {
-            key: 'name',
-            label: 'Name',
-            render: row => row.to
-                ? <Link className="table-name-link" to={row.to}>{row.name}</Link>
-                : row.name
-        },
-        { key: 'winPct', label: 'Win %', render: row => `${row.winPct}%` },
-        { key: 'record', label: 'W : L' },
-    ];
+    const Rankings = (event) => (
+        <ListGroup as="ol" numbered>
+            {event.filter(r => r[1].wins + r[1].losses > 4).slice(0, 10).map((r, i) => (
+                <ListGroup.Item key={`${r[1].name}-${i}`}>
+                    <Row>
+                        <Col xs={6}>{i + 1}. {r[1].name}</Col>
+                        <Col xs={6}>{r[1].percentage}% (W: {r[1].wins}, L: {r[1].losses})</Col>
+                    </Row>
+                </ListGroup.Item>
+            ))}
+        </ListGroup>
+    );
 
     return (
-        <PageShell title="Team Statistics" eyebrow="Current rankings by win %">
-            <div className="toolbar-row">
-                <TabControl tabs={EVENT_TABS} active={activeEvent} onChange={setActiveEvent} />
-                <SeasonSelect value={seasonValue} onChange={handleSeasonChange} />
-            </div>
-            <RankingsTable columns={columns} rows={rankings} />
-        </PageShell>
-    )
-}
+        <>
+            <Container>
+                <Row className="page-title">
+                    Welcome to the official site of the Waterloo Warriors Varsity Badminton Team Statistics!
+                </Row>
+                <SeasonSelector
+                    setStart={(start) => setSeasonStart(start)}
+                    setEnd={(end) => setSeasonEnd(end)}
+                />
+                <Row className="page-title">
+                    Current Rankings By Win %:
+                </Row>
+                <Container>
+                    <Row>
+                        <Col sm={12} md={6}>
+                            <div className="table-header">Singles Rankings</div>
+                            {Rankings(singlesRankings)}
+                        </Col>
+                        <Col sm={12} md={6}>
+                            <div className="table-header">Doubles Rankings</div>
+                            {Rankings(doublesRankings)}
+                        </Col>
+                        <Col sm={12} md={6}>
+                            <div className="table-header">Mixed Rankings</div>
+                            {Rankings(mixedRankings)}
+                        </Col>
+                    </Row>
+                </Container>
+            </Container>
+        </>
+    );
+};
 
-export default LandingPage
+export default LandingPage;
