+++ front-end/src/components/ReportMatch.js
@@ -1,33 +1,29 @@
-import { Form } from "react-bootstrap";
-import '../index.css';
-import { useState } from "react";
+import { Container, Tabs, Tab, Row, Col, Form } from "react-bootstrap";
 import SinglesForm from "../Forms/SinglesForm";
 import DoublesForm from "../Forms/DoublesForm";
 import MixedForm from "../Forms/MixedForm";
-import { PageShell, TabControl } from "./RedesignUI";
 
-const reportTabs = [
-    { key: 'Singles', label: 'Singles' },
-    { key: 'Doubles', label: 'Doubles' },
-    { key: 'Mixed', label: 'Mixed' },
-];
-
-const ReportMatchComponent = () => {
-    const [activeTab, setActiveTab] = useState('Singles');
-    const ActiveForm = activeTab === 'Singles'
-        ? SinglesForm
-        : activeTab === 'Doubles'
-            ? DoublesForm
-            : MixedForm;
-
-    return (
-        <PageShell title="Report Match" className="report-page">
+const ReportMatchComponent = () => (
+    <>
+        <Container>
+            <Row>
+                <Col><p className="page-title">REPORT MATCH</p></Col>
+            </Row>
             <Form id="match-form">
-                <TabControl tabs={reportTabs} active={activeTab} onChange={setActiveTab} className="report-tabs" />
-                <ActiveForm />
+                <Tabs defaultActiveKey="Singles">
+                    <Tab eventKey="Singles" title="Singles" tabClassName="report-tab">
+                        <SinglesForm />
+                    </Tab>
+                    <Tab eventKey="Doubles" title="Doubles" tabClassName="report-tab">
+                        <DoublesForm />
+                    </Tab>
+                    <Tab eventKey="Mixed" title="Mixed" tabClassName="report-tab">
+                        <MixedForm />
+                    </Tab>
+                </Tabs>
             </Form>
-        </PageShell>
-    )
-}
+        </Container>
+    </>
+);
 
 export default ReportMatchComponent;
